refactor: address review — merge into HubExtensions, add SentrySdk facade, configurable scope
- Move RecordTransaction into HubExtensions.cs (drop the separate file).
- Add SentrySdk.RecordTransaction facade so callers without an IHub
  instance can use it.
- Expose the fresh capture scope via ITransactionRecorder.ConfigureScope,
  so data captured alongside the original trace (user/tags/contexts) can
  be set without inheriting the current live scope.
